Added mypy and basedpyright to the dev dependencies.
The `returns` package only truly works with the `mypy` type checker,
as they provide a special plugin for it. The basedpyright type checker, however,
seems to also work decently well. basedpyright also provides many IDE
features as a language server, which is why I prefer to use it for this
project. For completeness sake, both are added to the dependencies, but
basedpyright is preferred.